Add regression test for #3112, improve and relax wildcard support
